+++ App.tsx
@@ -3,7 +3,7 @@ import AuthGate from './components/AuthGate';
 import DashboardApp from './DashboardApp';
 
 const App: React.FC = () => {
-  return <AuthGate>{({ session }) => <DashboardApp session={session} />}</AuthGate>;
+  return <AuthGate>{({ session, access }) => <DashboardApp session={session} access={access} />}</AuthGate>;
 };
 
 export default App;